frontend deploy

+++ backend/db/migrations/20241108185118-create-spot.js
@@ -54,7 +54,7 @@ module.exports = {
           allowNull: false,
         },
         price: {
-          type: Sequelize.NUMERIC,
+          type: Sequelize.FLOAT,
           allowNull: false,
         },
         avgRating: {
